Add new filter widget/component for generic filtering
A new generic filter widget (FilterPresenterWidget) was added.  FilterPresenterWidget. FilterPresenterWidget can be dropped in any Presenter. FilterPresenterWidget take a list of instances of a subclass of FilterItemPresenterWidget and displays them in a horizontal form.  Filters are set in a modal popup. Currently support for TextBox, DropDown and Typeahead FilterWidgets were added.