Feature: Add a dynamic favicons which show unread notifications
A custom favicon was added to the site. In additional to this a small js library
(http://lab.ejci.net/favico.js/) that allows to dynamically add badges to the favicon and
a simple GWT wrapper (Favico.java and FaviconOptions.java) were added.
The favicon now shows the same badge as the My Account menu.