Feature: Candidate gene list enrichment analysis added
Now it is possible to run candidate gene list enrichment analysis for a set of
gwas-results or a set of candidate gene lists. The analysis is done by sending a
message to the message-broker. The celery worker will run it. Furthermore it is
now possible to display the list of enrichments on the Experiment/Phenotype
and Study level.