Bugfix #8: Categegories display counts and support for loading indicator
The categories of the global search now display the icon of each category and
also the counts/totalcounts. Furthermore a loading indicator is displayed when
the query takes longer.
The category title is a link that opens the corresponding list views with the search
term as a query URL parameter.