Bugfix: Avoid throwing Exception and displaying default error page
When the User navigates to the #/meta/gene view by default no Gene is selected.
This will cause an exception in the RestProviderController.getGEnomeStatsData
method which will lead to an Javascript Exception on the client side because
GeneViewer can't deal with it. Instead of throwing an Exception, the Exception will
be logged and an empty DataTable will sent back.

This closes #45