Enhancement: Replace MD5PasswordEncoder with BcryptPasswordEncoder
MD5 is not secure enough. Current best practice is to use bcrypt. In order to
migrate the users from md5 to bcrypt a new MigrateUserPasswordEncoder
was added that will check first if the password is encoded with the md5
encoder. if so it will re-encode it using bcrypt and save it in the database. The
MigrateUserPasswordEncoder will be used until all passwords are migrated.

All other services (change password, reset forgotten password and
registernewUser) now use the BcryptPasswordEncoder

This closes #64