Enhancement: Add support for batch upload of phenotypes
This is a big refactor of the PhenotypeUploadWizardPresenterWidget to allow
batch upload of phenotypes. This enables us collapse the the ISA-TAB and CSV
functionality. From now on there is only one phenotype upload functionality
(PhenotypeUploadWizardPresenterWidget) that is used no matter if the user
uploads a CSV or a ISA-TAB archive. All uploaded values will be treated as
MEASURE.

The new PhenotypeUploadWizardPresenterWidget now not only shows a
table with the values but also display various statistics (Histogram, etc) and
has proper validation checks. Some Tests were also added to IsaTabExporer
to make sure that the code is properly covered. This closes #68