Enhancement: Encapsulate Elasticsearch's searching/indexing functionality
EsDocument is a marker interface for any POJO or JPA Entity which is used to
find out the mapping name of the corresponding document in ES.

EsIndexer encapsulates the indexing behavior for any generic EsDocument
annotated entity.

 EsIndexerApp is a command line app that allows to bulk index an entire entity.
 It can be used instead of a JDBC river.

EsSearcher encapsulates the functionality for searching for documents in ES. It uses Aggregations instead of Facets (#52) and filtered query instead of post-filter (#53). In addition to All, Public, Private documents also Shared document can be fetched.