Enhancement: Add support for downloading Manhattan GWAS plots
The backend now supports two new RestEndpoints to generate and
download Manhattan plots. The backend will call a docker container that uses
Python to generate the figures and then serves it to the user. Furthermore a
new PlotDownloadPopup Widget was added to allow the user to specify
Chromosome, Format and MAC filter. Generating can be done for
both normal analysis and also gwasviewer uploads.
This closes 81.