Enhancement: Update SNP Annotation to the latest version
Update the SNP Annotation in the Elasticsearch index to the latest version.
The ES template for the snps document was changed to match the information
that are provided by the snpeff output format. Furthermore the import script
was adapted to read a custom snpeff format file and index the information in
ES.

The SNPDetailView/Presenter will now display those additional information
(effect, function, amino acid change, etc). The corresponding DTOs were also
updated and renamed for clarity. This closes #82