Enhancement: Added support for viewing SNPs
A new SNPViewerPresenter was added that allows users to view snps in a
specific region for a specific genotype. The SNPs can be filtered by Phenotypes.
The functionality of SNPDetailPresenter was moved to a seperate
SNPDetailPresenterWidget which is used in SNPDetailPresenter and
SNPViewerPresenter to display information about a specific SNP. Futhermore
the genotype section contains a left hand side menu with two options: 1.)
SNPViewer and 2.) Genome Browser.