Bugfix: Disable paging controls when request is executing
The CustomPager has a new method setPageControlDisabled to disable/enable
 paging controls.
To avoid a Request is already in progress error the paging controls are disabled
and re-enabled once the request is finished.
In addition to this the in the MetaAnalysisGenePresenter the context is not
stored as a class variable but locally inside the onRangeChanged callback.

 This fixes #100.