Changed mapping for ES 2.x upgrade (#123)
Removed the doc_format from the mapping because in ES 2.x it is the default.
Removed boost settings from the mapping because best practice is to
provide it during search time.
Added search_analyzer for ngram fields becaue ES 2.x by default uses the
same analyzer for indexing and search.