Bugfix: Improve performance when switching between facets
Transferring a RequestFactory entity with many relationships (i.e.
study/candidate gene list in the context of enrichment) takes very long
because for some reason all properties of the object graph have to be
traversed (client side). This takes a couple of seconds.

So instead of sending an entity only the id + an enum is sent to the backend.
This improves performance quite a bit. Futhermore some additonal
bugfixes were done.
This fixes #145 and #146