Bugfix: Implement workaround for http issue with google auth
Inject a custom GoogleOauth2Template that replaces http with https.
This is required due to recent change in Google OAUTH2 which
requires https for non localhost redirect urls.
Spring security social > 2.x would make this workaround obsolete
as it detects that the app is running behind a proxy that does TLS termination.